Add delete fucntion
